+++ totalRP3/core/impl/globals.lua
@@ -47,7 +47,7 @@ TRP3_API.globals = {
 	addon_name_me = "Total RP 3",
 	addon_id_length = 15,
 
-	version = 39,
+	version = 40,
 
 	--@debug@
 	version_display = "-dev",